# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /api
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /signIn:
    x-swagger-router-controller: account
    post:
      description: Signin account with public key and signature
      operationId: signIn
      parameters:
        - name: body
          in: body
          description: parameter body with public key and signature
          schema:
            $ref: '#/definitions/SignIn'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /signUp:
    x-swagger-router-controller: account
    post:
      description: Create account by publicKey, require transaction be created by client
      operationId: signUp
      parameters:
        - name: body
          in: body
          description: parameter with transaction be created by client
          schema:
            $ref: '#/definitions/PostTransaction'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /account/{publicKey}:
    x-swagger-router-controller: account
    get:
      description: get account info
      operationId: profile
      parameters:
        - name: publicKey
          in: path
          description: publicKey
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        
    post:
      description: Update account require transaction to commit
      operationId: update
      parameters:
        - name: publicKey
          in: path
          description: publicKey
          required: true
          type: string
        - name: body
          in: body
          description: parameter with transaction be created by client
          schema:
            $ref: '#/definitions/PostTransaction'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        
  /account/{publicKey}/follow:
    x-swagger-router-controller: account
    get:
      description: get follow in account
      operationId: getFollow
      parameters:
        - name: publicKey
          in: path
          description: publicKey
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        
    post:
      description: follow orther account require transaction to commit
      operationId: follow
      parameters:
        - name: publicKey
          in: path
          description: publicKey
          required: true
          type: string
        - name: body
          in: body
          description: parameter with transaction be created by client
          schema:
            $ref: '#/definitions/PostTransaction'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        
  /tweet:
    x-swagger-router-controller: tweet
    get:
      description: get all tweet
      operationId: getAll
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        
    post:
      description: Post new tweet
      # used as the method name of the controller
      operationId: post
      parameters:
        - name: body
          in: body
          description: parameter with transaction be created by client
          schema:
            $ref: '#/definitions/PostTransaction'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        
  /tweet/{tweetId}:
    # binds a127 app logic to a route
    x-swagger-router-controller: tweet
    get:
      description: get all interact in this tweet
      # used as the method name of the controller
      operationId: getOne
      parameters:
        - name: tweetId
          in: path
          description: id of topic
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
  /tweet/{tweetId}/interact:
    # binds a127 app logic to a route
    x-swagger-router-controller: tweet
    post:
      description: post new interact in this tweet
      operationId: interact
      parameters:
        - name: tweetId
          in: path
          description: id of topic
          required: true
          type: string
        - name: body
          in: body
          description: parameter with transaction be created by client
          schema:
            $ref: '#/definitions/PostTransaction'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        

  /payment/{publicKey}:
    # binds a127 app logic to a route
    x-swagger-router-controller: payment
    get:
      description: get balance in my account
      # used as the method name of the controller
      operationId: balance
      parameters:
        - name: publicKey
          in: path
          description: publicKey
          required: true
          type: string
        - name: body
          in: body
          description: parameter with public key
          schema:
            $ref: '#/definitions/CreateAccount'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        
    post:
      description: send payment to orther account
      # used as the method name of the controller
      operationId: send
      parameters:
        - name: publicKey
          in: path
          description: publicKey
          required: true
          type: string
        - name: body
          in: body
          description: parameter with transaction be created by client
          schema:
            $ref: '#/definitions/PostTransaction'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"           
  /payment/history:
    # binds a127 app logic to a route
    x-swagger-router-controller: payment
    get:
      description: Get all history payment
      # used as the method name of the controller
      operationId: history
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  CreateAccount:
    type: object
    required:
      - transaction
    properties:
      transaction:
        type: string
  PostTransaction:
    type: object
    required:
      - txs
    properties:
      txs:
        type: string
  SignIn:
    type: object
    required:
      - publicKey
      - signature
    properties:
      publicKey:
        type: string
      signature:
        type: string
  MessageResponse:
    properties:
      statusCode:
        type: integer
      message:
        type: string
      value:
        type: object
  ErrorResponse:
    required:
      - message
    properties:
      statusCode:
        type: integer
      message:
        type: string